requirements.txt
livekit
python-dotenv
numpy
scipy
tensorflow
pytorch
